E-Commerce_Springboot_Application

+++ src/main/java/com/example/e_commerce/model/Product.java
@@ -21,7 +21,7 @@ public class Product {
     private String description;
     private BigDecimal price;
     private String brand;
-    private int inventory;
+    private Long inventory;
 
     @ManyToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "Category_Id")
